Rewrote the stock quote caching system. Previously each task in Venice created an instance of QuoteCache and used that when reading stock quotes. The problem with this approach is that if different tasks wanted to play with the same data, it was loaded twice and kept in memory twice.
To solve this problem, QuoteCache was made into a singleton. Its new task
was to provide a single repository of stock quotes.

Now each task creates a QuoteBundle instance which describes which stock
quotes they are interested in. The QuoteBundle talks to the QuoteCache
to retrieve the quotes.

The new QuoteBundleCache performs caching based on the QuoteBundle level.