Added simplify() method which can simplify expressions. This is (hopefully) beneficial to the genetic programme code which can create some pretty horrific statements such as:
((held > 5 or true) and (order == order))

Which is now simplified to "true". This is important because now the
GP will realise the expression is merely the "true" statement so it
won't waste our time trying to paper trade with it for the 1000th time.

As a result of doing this work I implemented the equals() method
which can tell whether two expression trees are equal.

E.g.

(a + b) is the same as (b + a)
lag(day_close, 0, 0) is the same as (lag_day_close, 0, 0)

This way I can simplify things like (a*b) == (a*b) and
(a and b) or (b and a) to true.