Replaced use of float with double to increase precision. This was especially necessary for Gondola as I was using floats to represent integers. I'm now using doubles which should work much better.
I created a new class - OrderCache - which is used when paper trading.
It caches the order of symbols for each day. This operation took a long,
long time and it speeds everything up. It is also means that you can
now apply an order and it is only slightly slower than having no
order.