Changed expression tree so that it is no longer based on DefaultMutableTreeNode. I did this mainly to increase speed in the GP which relies heavily on creating these classes. The DefaultMutableTreeNode uses a Vector to hold and manipulate child entries, which is overkill in this structure. Also it has a variety of other fields which are unncessary. This speeded up the GP by about 20%.
Another reason for doing this was to reduce the memory required by
storing equations.

And finally it also paves the way to remove the parent field from the
nodes. This would enable the quote expression leaf nodes to be
canonical. It's unclear at this point in time whether there is really
any big gain in doing that.

Also as part of the float -> double change, I changed constants such
as 0.0F -> 0.0D.