I fixed a couple of problems Venice was having under Java 1.5beta:
1) I was asking the JDesktopPane for the DesktopManager and trying to
cast this back to my DesktopManager. This worked in the past, but does
not work with Java's new Gnome L&F. I presume the Gnome L&F has some
kind of wrapper around my DesktopManager. I now pass a pointer to my
DesktopManager and get a pointer to the JDesktopPane which will not
suffer from this casting problem.

2) When graphing a symbol, for some reason the "graph" symbol can be
NULL, causing a NullPointerException when the window's title is
computed. I don't really understand why, but since this module will
be upgraded soon, I just worked around it and it works fine.