Merged a bug fix from Bryan Lin.
What was happening was that when the equation results were displayed, it
was using the symbol to retrieve the equation result. The problem was
that if you had quotes on different days for a single symbol, it would ignore
the dates, and just use the symbol to look it up. Now it uses
both the symbol AND the date.