I've got the unit tests working again!
There was a problem with an earlier version of JUnit/Ant were they
simply did not play well together and I had to abandom the tests.
It looks like this has all been fixed now.

There were a few minor interface changes that required modifications to
build the tests, and overall all the tests still worked. The only
problem was that ClauseExpression was converting equations slightly
differently so that made a couple of tests needlessly fail.

I've also fixed a problem were due to the ClauseExpression change, code
was being printed like:

{ true }

instead of:

true

I've modified ClauseExpression so that if the clause only contains
a single expression, it drops the {}'s.