Fixed two bugs with import/export. The first was that values were being written out in the form "$2,500" which conflicted with the user of commas as item separators. So I have dropped use of commas for separators in favour of tabs.
Also when writing transactions which did not contain a symbol, the
text "null" was written. While this was acceptable, unfortunately
when the text was being imported, it would interpret this to be a
symbol named "NULL" (!) This did not seem to adversely affect Venice
but an empty symbol field is now written "-" to avoid this potential
error.

To ensure that portfolio import/export always works, I have written
an extensive unit test of the portfolio code. To do this I had
to refactor the portfolio code slightly (moving the read/write
methods from PortfolioModule.java to Portfolio.java) and also
added various comparision methods for testing purposes.

Venice needs more of these regression tests.