Code cleanup.
I moved the SQL date string format code to DatabaseQuoteSource so
that the generic toString() method would, by default, print a sensible
date format (not the SQL format!). This simplified the AbstractTable
code as it could treat a date like any other object. It should also
dramatically improve the time request to format a date. This might have
a slight impact on table redraw times.